Implement ownership-based authorization

+++ src/main/java/org/example/soundly/serviceIMP/SongServiceIMP.java
@@ -1,12 +1,18 @@
 package org.example.soundly.serviceIMP;
 
 import lombok.RequiredArgsConstructor;
+import org.example.soundly.Enum.Role;
 import org.example.soundly.dto.SongResponse;
 import org.example.soundly.entity.Song;
 import org.example.soundly.repository.SongRepository;
 import org.example.soundly.service.SongService;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.example.soundly.entity.User;
+import org.example.soundly.repository.UserRepository;
 
 import java.io.File;
 import java.io.IOException;
@@ -16,11 +22,20 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class SongServiceIMP implements SongService {
+
     private final SongRepository songRepository;
 
+    private final UserRepository userRepository;
+
     @Override
-    public String uploadSong(String title, String artist, String genre, MultipartFile file) {
-        try{
+    public String uploadSong(
+            String title,
+            String artist,
+            String genre,
+            MultipartFile file) {
+
+        try {
+
             String uploadDir =
                     System.getProperty("user.dir")
                             + File.separator
@@ -31,31 +46,51 @@ public class SongServiceIMP implements SongService {
 
             File directory = new File(uploadDir);
 
-            if(!directory.exists()){
+            if (!directory.exists()) {
                 directory.mkdirs();
             }
-            String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
+
+            String fileName =
+                    System.currentTimeMillis()
+                            + "_"
+                            + file.getOriginalFilename();
 
             String filePath = uploadDir + fileName;
 
-
-
-
             file.transferTo(new File(filePath));
 
+            // Get currently logged-in user
+            Authentication authentication =
+                    SecurityContextHolder
+                            .getContext()
+                            .getAuthentication();
 
+            String email = authentication.getName();
 
-            Song song = Song.builder().title(title).artist(artist).genre(genre).filePath(filePath).build();
+            User user = userRepository
+                    .findByEmail(email)
+                    .orElseThrow(() ->
+                            new RuntimeException("User not found"));
 
+            Song song = Song.builder()
+                    .title(title)
+                    .artist(artist)
+                    .genre(genre)
+                    .filePath(filePath)
+                    .uploader(user)
+                    .build();
 
             songRepository.save(song);
 
-
             return "Song uploaded successfully";
 
-        }catch (IOException e){
+        } catch (IOException e) {
+
             e.printStackTrace();
-            throw new RuntimeException("Failed to upload song", e);
+
+            throw new RuntimeException(
+                    "Failed to upload song",e
+            );
         }
     }
 
@@ -78,16 +113,42 @@ public class SongServiceIMP implements SongService {
     }
 
     @Override
-    public void deleteSong(long id){
+    public String deleteSong(Long songId){
 
-        Song song = songRepository.findById(id).orElseThrow(() -> new RuntimeException("Song not found"));
+        Song song = songRepository
+                .findById(songId)
+                .orElseThrow();
 
-        File file = new File(song.getFilePath());
+        Authentication authentication =
+                SecurityContextHolder.getContext()
+                        .getAuthentication();
 
-        if(file.exists()){
-            file.delete();
+        String email = authentication.getName();
+
+        User currentUser =
+                userRepository.findByEmail(email)
+                        .orElseThrow();
+
+        System.out.println("Current User ID: " + currentUser.getId());
+        System.out.println("Current User Role: " + currentUser.getRole());
+
+        System.out.println("Song Uploader: " + song.getUploader());
+
+        if(song.getUploader() != null){
+            System.out.println("Uploader ID: " + song.getUploader().getId());
         }
+
+        if(currentUser.getRole() != Role.ADMIN
+                &&
+                song.getUploader().getId() != currentUser.getId()) {
+
+            throw new RuntimeException(
+                    "You can only delete your own songs");
+        }
+
         songRepository.delete(song);
+
+        return "Song deleted successfully";
     }
 
     @Override
@@ -103,4 +164,22 @@ public class SongServiceIMP implements SongService {
                         .build())
                 .toList();
     }
+
+    @Override
+    public List<SongResponse> searchSongsByGenre(String genre){
+        return songRepository
+                .findByGenreContainingIgnoreCase(genre)
+                .stream()
+                .map(song -> SongResponse.builder()
+                        .id(song.getId())
+                        .title(song.getTitle())
+                        .artist(song.getArtist())
+                        .genre(song.getGenre())
+                        .build())
+                .toList();
+    }
+
+
+
+
 }
